last update

+++ uptable.sh
@@ -16,6 +16,7 @@ then
 c=$(grep -n "$att0" Databases/$DBconnect/"meta$tablename")
 echo $c
 cpk=$(echo $c|cut -f4 -d:)
+ctype=$(echo $c|cut -f3 -d:)
 cn=$(echo $c|cut -f1 -d:)
 (( cn++ ))
 if [[ $cpk == "Primary_Key" ]]
@@ -27,10 +28,23 @@ if [[ $flag != 1 ]]
 then
 att1=$(cut -f1 -d: Databases/$DBconnect/"meta$tablename"|grep -nw "$weratt"|cut -f1 -d:)
 (( att1++ ))
+if [[ $ctype == int ]]
+then
+if [[ $value =~ $int ]]
+then
 touch Databases/$DBconnect/temp
 awk -F: -v wat=$att1 -v at=$att2 -v v=$value -v wv=$wervalue '{OFS=FS}{if($wat==wv){$at=v;print $0 }else {print $0}}' Databases/$DBconnect/$tablename >> Databases/$DBconnect/temp
 cp Databases/$DBconnect/temp Databases/$DBconnect/$tablename
 rm Databases/$DBconnect/temp
+else
+zenity --error --title "error" --text "attribute value must be number"
+fi
+else
+touch Databases/$DBconnect/temp
+awk -F: -v wat=$att1 -v at=$att2 -v v=$value -v wv=$wervalue '{OFS=FS}{if($wat==wv){$at=v;print $0 }else {print $0}}' Databases/$DBconnect/$tablename >> Databases/$DBconnect/temp
+cp Databases/$DBconnect/temp Databases/$DBconnect/$tablename
+rm Databases/$DBconnect/temp
+fi
 else
 zenity --error --title "error" --text "attribute is PK \nDuplicate value"
 fi
